requirements.txt
moviepy==1.0.3
numpy==1.24.1
opencv_python==4.7.0.72
Pillow==9.5.0
pydub==0.25.1
tqdm==4.64.1
